+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/SubjectCredentialStore.kt
@@ -93,5 +93,4 @@ interface SubjectCredentialStore {
             override val schemaUri: String,
         ) : StoreEntry
     }
-
 }